requirements.txt
python-dotenv~=0.19.0
transformers~=4.41.2
pandas~=2.2.0
tqdm~=4.62.3
numpy~=1.23.5
datasets~=2.19.2
transformers[torch]
tensorboardX~=2.6.2.2
sentencepiece
protobuf